backend/requirements.txt
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
docstring_parser==0.16
fastapi==0.95.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
numpy==2.2.4
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==1.10.4
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.26.1
tqdm==4.67.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.22.0
vertexai==1.71.1
